% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN.r
\name{createTemplate}
\alias{createTemplate}
\title{Create Template}
\usage{
createTemplate(x, y)
}
\arguments{
\item{x}{SOME DATA}

\item{y}{SOME OTHER DATA}
}
\value{
The newly created template
}
\description{
Create Template
}
\examples{
newtemplate <- createTemplate(sampletemplate)
}
