% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIN.r, R/app.R
\docType{package}
\name{EmailTemplate}
\alias{EmailTemplate}
\title{Create Email Template}
\usage{
EmailTemplate()
}
\description{
This package allows you to run a shiny app that helps you to save and load email templates.
}
\section{run app}{

\emph{EmailTemplate} runs app.
}

